@@ -1,12 +1,10 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>io.github.universal-variability-language</groupId>
     <artifactId>fm-metamodel</artifactId>
-    <version>1.0</version>
+    <version>1.1</version>
 
     <name>UVL Java Metamodel</name>
     <description>This project host a java feature model metamodel for UVL </description>
@@ -47,10 +45,11 @@
     </developers>
 
     <scm><!-- TODO-->
-        <connection>scm:git:git@github.com:Universal-Variability-Language/java-uvl-metamodel.git</connection>
-        <developerConnection>scm:git:git@github.com:Universal-Variability-Language/java-uvl-metamodel.git</developerConnection>
-        <url>https://github.com/Universal-Variability-Language/java-uvl-metamodel</url>
-    </scm>
+        <connection>scm:git:git@github.com:Universal-Variability-Language/java-fm-metamodel.git</connection>
+        <developerConnection>scm:git:git@github.com:Universal-Variability-Language/java-fm-metamodel.git</developerConnection>
+        <url>https://github.com/Universal-Variability-Language/java-fm-metamodel</url>
+      <tag>HEAD</tag>
+  </scm>
 
     <distributionManagement>
         <snapshotRepository>
@@ -65,8 +64,8 @@
 
     <properties>
         <antlr4.version>4.13.1</antlr4.version>
-        <maven.compiler.source>8</maven.compiler.source>
-        <maven.compiler.target>8</maven.compiler.target>
+        <maven.compiler.source>11</maven.compiler.source>
+        <maven.compiler.target>11</maven.compiler.target>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
     </properties>
