@@ -3,6 +3,8 @@ package de.vill.conversion;
 import com.google.common.collect.Sets;
 import de.vill.model.*;
 import de.vill.model.constraint.*;
+import de.vill.model.expression.AggregateFunctionExpression;
+import de.vill.model.expression.BinaryExpression;
 import de.vill.model.expression.Expression;
 import de.vill.model.expression.LiteralExpression;
 
@@ -23,7 +25,7 @@ public class ConvertSMTLevel implements IConversionStrategy {
     public void convertFeatureModel(FeatureModel rootFeatureModel, FeatureModel featureModel) {
         List<Constraint> constraints = featureModel.getFeatureConstraints();
         constraints.addAll(featureModel.getOwnConstraints());
-        constraints.stream().forEach(x -> replaceEquationInConstraint(x));
+        constraints.stream().forEach(this::replaceEquationInConstraint);
         List<Constraint> replacements = new LinkedList<>();
         for (Constraint constraint : featureModel.getOwnConstraints()) {
             if (constraint instanceof ExpressionConstraint) {
@@ -69,8 +71,13 @@ public class ConvertSMTLevel implements IConversionStrategy {
     private Set<Feature> getFeaturesInExpression(Expression expression) {
         Set<Feature> featuresInEquation = new HashSet<>();
         if (expression instanceof LiteralExpression) {
-            featuresInEquation.add(((LiteralExpression) expression).getFeature());
-        } else {
+            LiteralExpression literalExpression = (LiteralExpression) expression;
+            if (literalExpression.getContent() instanceof Feature) {
+                featuresInEquation.add((Feature) ((LiteralExpression) expression).getContent());
+            } else if (literalExpression.getContent() instanceof Attribute<?>) {
+                featuresInEquation.add(((Attribute<?>) literalExpression.getContent()).getFeature());
+            }
+        }  else {
             for (Expression subExpression : expression.getExpressionSubParts()) {
                 featuresInEquation.addAll(getFeaturesInExpression(subExpression));
             }
@@ -93,8 +100,7 @@ public class ConvertSMTLevel implements IConversionStrategy {
             feature = allFeatures.iterator().next();
             allFeatures.remove(feature);
         }
-        Constraint literalConstraint = new LiteralConstraint(feature.getFeatureName());
-        ((LiteralConstraint) literalConstraint).setFeature(feature);
+        Constraint literalConstraint = new LiteralConstraint(feature);
         if (!selectedFeatures.contains(feature)) {
             literalConstraint = new NotConstraint(literalConstraint);
         }
@@ -128,7 +134,7 @@ public class ConvertSMTLevel implements IConversionStrategy {
         if (attributeConstraint != null) {
             if (attributeConstraint.getValue() instanceof ExpressionConstraint) {
                 Constraint equationReplacement = convertEquationToConstraint((ExpressionConstraint) attributeConstraint.getValue());
-                feature.getAttributes().put("constraint", new Attribute<>("constraint", equationReplacement));
+                feature.getAttributes().put("constraint", new Attribute<>("constraint", equationReplacement, feature));
             }
         }
         if (attributeConstraintList != null && attributeConstraintList.getValue() instanceof List<?>) {
@@ -141,7 +147,7 @@ public class ConvertSMTLevel implements IConversionStrategy {
                     newConstraintList.add(constraint);
                 }
             }
-            feature.getAttributes().put("constraints", new Attribute<>("constraints", newConstraintList));
+            feature.getAttributes().put("constraints", new Attribute<>("constraints", newConstraintList, feature));
         }
     }
 
