@@ -4,6 +4,8 @@ import de.vill.model.Feature;
 import de.vill.model.FeatureModel;
 import de.vill.model.Group;
 import de.vill.model.LanguageLevel;
+import de.vill.util.Constants;
+
 import java.util.Collections;
 import java.util.HashSet;
 import java.util.Set;
@@ -26,8 +28,8 @@ public class DropTypeLevel implements IConversionStrategy {
 
     private void traverseFeatures(final Feature feature) {
         feature.setFeatureType(null);
-        feature.getAttributes().remove("type_level_value");
-        feature.getAttributes().remove("type_level_value_length");
+        feature.getAttributes().remove(Constants.TYPE_LEVEL_VALUE);
+        feature.getAttributes().remove(Constants.TYPE_LEVEL_LENGTH);
 
         for (final Group group : feature.getChildren()) {
             for (final Feature subFeature : group.getFeatures()) {
