@@ -1,5 +1,8 @@
 package de.vill.model.constraint;
 
+import de.vill.model.building.VariableReference;
+
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
@@ -23,11 +26,9 @@ public class AndConstraint extends Constraint {
 
     @Override
     public String toString(boolean withSubmodels, String currentAlias) {
-        StringBuilder result = new StringBuilder();
-        result.append(left.toString(withSubmodels, currentAlias));
-        result.append(" & ");
-        result.append(right.toString(withSubmodels, currentAlias));
-        return result.toString();
+        return  left.toString(withSubmodels, currentAlias) +
+                " & " +
+                right.toString(withSubmodels, currentAlias);
     }
 
     @Override
@@ -67,5 +68,13 @@ public class AndConstraint extends Constraint {
         }
         AndConstraint other = (AndConstraint) obj;
         return Objects.equals(left, other.left) && Objects.equals(right, other.right);
+    }
+
+    @Override
+    public List<VariableReference> getReferences() {
+        List<VariableReference> references = new ArrayList<>();
+        references.addAll(left.getReferences());
+        references.addAll(right.getReferences());
+        return references;
     }
 }
