@@ -1,7 +1,9 @@
 package de.vill.model.constraint;
 
+import de.vill.model.building.VariableReference;
 import de.vill.model.expression.Expression;
-import java.util.Collections;
+
+import java.util.ArrayList;
 import java.util.List;
 
 public class EqualEquationConstraint extends ExpressionConstraint {
@@ -15,12 +17,15 @@ public class EqualEquationConstraint extends ExpressionConstraint {
     }
 
     @Override
-    public List<Constraint> getConstraintSubParts() {
-        return Collections.emptyList();
+    public Constraint clone() {
+        return new EqualEquationConstraint(this.left, this.right);
     }
 
     @Override
-    public Constraint clone() {
-        return new EqualEquationConstraint(this.left, this.right);
+    public List<VariableReference> getReferences() {
+        List<VariableReference> references = new ArrayList<>();
+        references.addAll(left.getReferences());
+        references.addAll(right.getReferences());
+        return references;
     }
 }
