@@ -1,5 +1,7 @@
 package de.vill.model.constraint;
 
+import de.vill.model.building.VariableReference;
+
 import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
@@ -19,7 +21,13 @@ public class NotConstraint extends Constraint {
     public String toString(boolean withSubmodels, String currentAlias) {
         StringBuilder result = new StringBuilder();
         result.append("!");
-        result.append(content.toString(withSubmodels, currentAlias));
+        if (content instanceof VariableReference || content instanceof ParenthesisConstraint) {
+            result.append(content.toString(withSubmodels, currentAlias));
+        } else {
+            result.append("(");
+            result.append(content.toString(withSubmodels, currentAlias));
+            result.append(")");
+        }
         return result.toString();
     }
 
@@ -55,5 +63,10 @@ public class NotConstraint extends Constraint {
         }
         NotConstraint other = (NotConstraint) obj;
         return Objects.equals(content, other.content);
+    }
+
+    @Override
+    public List<VariableReference> getReferences() {
+        return content.getReferences();
     }
 }
