@@ -1,19 +1,28 @@
 package de.vill.model.expression;
 
 import de.vill.model.Feature;
+import de.vill.model.building.VariableReference;
 import de.vill.util.Constants;
-import java.util.Arrays;
-import java.util.List;
-import java.util.Objects;
-import java.util.Set;
 
-public class AddExpression extends Expression {
+import java.util.*;
+
+public class AddExpression extends BinaryExpression {
     private Expression left;
     private Expression right;
 
     public AddExpression(Expression left, Expression right) {
         this.left = left;
         this.right = right;
+    }
+
+    @Override
+    public Expression getLeft() {
+        return left;
+    }
+
+    @Override
+    public Expression getRight() {
+        return right;
     }
 
     @Override
@@ -44,20 +53,7 @@ public class AddExpression extends Expression {
 
     @Override
     public double evaluate(Set<Feature> selectedFeatures) {
-        double leftResult;
-        if (left instanceof LiteralExpression && !selectedFeatures.contains(((LiteralExpression) left).getFeature())) {
-            leftResult = 0;
-        } else {
-            leftResult = left.evaluate(selectedFeatures);
-        }
-        double rightResult;
-        if (right instanceof LiteralExpression
-            && !selectedFeatures.contains(((LiteralExpression) right).getFeature())) {
-            rightResult = 0;
-        } else {
-            rightResult = right.evaluate(selectedFeatures);
-        }
-        return leftResult + rightResult;
+        return left.evaluate(selectedFeatures) + right.evaluate(selectedFeatures);
     }
 
     @Override
@@ -78,5 +74,13 @@ public class AddExpression extends Expression {
         }
         AddExpression other = (AddExpression) obj;
         return Objects.equals(left, other.left) && Objects.equals(right, other.right);
+    }
+
+    @Override
+    public List<VariableReference> getReferences() {
+        List<VariableReference> references = new ArrayList<>();
+        references.addAll(left.getReferences());
+        references.addAll(right.getReferences());
+        return references;
     }
 }
