@@ -1,16 +1,20 @@
 package de.vill.model.expression;
 
+import de.vill.model.Attribute;
+import de.vill.model.Feature;
+import de.vill.model.GlobalAttribute;
+import de.vill.model.building.VariableReference;
 import de.vill.util.Constants;
 import java.util.Collections;
 import java.util.List;
 
 public class AvgAggregateFunctionExpression extends AggregateFunctionExpression {
-    public AvgAggregateFunctionExpression(String attributeName) {
-        super(attributeName);
+    public AvgAggregateFunctionExpression(GlobalAttribute reference) {
+        super(reference);
     }
 
-    public AvgAggregateFunctionExpression(String rootFeatureName, String attributeName) {
-        super(rootFeatureName, attributeName);
+    public AvgAggregateFunctionExpression(GlobalAttribute reference, Feature rootFeature) {
+        super(reference, rootFeature);
     }
 
     @Override
