@@ -1,29 +1,39 @@
 package de.vill.model.expression;
 
+import com.google.errorprone.annotations.Var;
+import de.vill.model.Attribute;
 import de.vill.model.Feature;
+import de.vill.model.building.VariableReference;
 import de.vill.util.Constants;
-import java.util.Arrays;
-import java.util.List;
-import java.util.Optional;
-import java.util.Set;
+import de.vill.util.Util;
 
-public class LengthAggregateFunctionExpression extends AggregateFunctionExpression {
-    private final String featureName;
+import java.util.*;
 
-    public LengthAggregateFunctionExpression(final String featureName) {
-        super(featureName, null);
-        this.featureName = featureName;
+public class LengthAggregateFunctionExpression extends Expression {
+
+    private VariableReference reference;
+
+    public LengthAggregateFunctionExpression(final VariableReference reference) {
+        this.reference = reference;
     }
 
     @Override
     public String toString(final boolean withSubmodels, final String currentAlias) {
-        return super.toString(withSubmodels, "len", currentAlias);
+        return String.format("len(%s)", Util.addNecessaryQuotes(reference.getIdentifier()));
     }
 
     @Override
     public List<Expression> getExpressionSubParts() {
         return Arrays.asList();
     }
+
+    @Override
+    public void replaceExpressionSubPart(Expression oldSubExpression, Expression newSubExpression) {
+    }
+
+    public VariableReference getReference() {return reference;}
+
+    public void setReference(VariableReference reference) {this.reference = reference;}
 
     @Override
     public String getReturnType() {
@@ -33,7 +43,7 @@ public class LengthAggregateFunctionExpression extends AggregateFunctionExpressi
     @Override
     public double evaluate(final Set<Feature> selectedFeatures) {
         final Optional<Feature> feature = selectedFeatures.stream()
-            .filter(f -> f.getFeatureName().equals(this.featureName))
+            .filter(f -> f.getFeatureName().equals(reference.getIdentifier())) // TODO: Is this correct?
             .findFirst();
 
         if (feature.isPresent()) {
@@ -45,5 +55,37 @@ public class LengthAggregateFunctionExpression extends AggregateFunctionExpressi
         }
 
         return 0;
+    }
+
+
+    @Override
+    public int hashCode() {
+        final int prime = 31;
+        int result = super.hashCode();
+        result = prime * result + Objects.hash(reference);
+        return result;
+    }
+
+    @Override
+    public int hashCode(int level) {
+        return 31 * level + (reference == null ? 0 : reference.getIdentifier().hashCode());
+    }
+
+
+    @Override
+    public boolean equals(Object obj) {
+        if (this == obj) {
+            return true;
+        }
+        if (obj == null || getClass() != obj.getClass()) {
+            return false;
+        }
+        LengthAggregateFunctionExpression other = (LengthAggregateFunctionExpression) obj;
+        return Objects.equals(reference, other.reference);
+    }
+
+    @Override
+    public List<VariableReference> getReferences() {
+        return List.of();
     }
 }
