@@ -1,15 +1,20 @@
 package de.vill.model.expression;
 
+import de.vill.model.Attribute;
+import de.vill.model.Feature;
+import de.vill.model.GlobalAttribute;
+import de.vill.model.building.VariableReference;
+
 import java.util.Arrays;
 import java.util.List;
 
 public class MaxAggregateFunctionExpression extends AggregateFunctionExpression {
-    public MaxAggregateFunctionExpression(String attributeName) {
-        super(attributeName);
+    public MaxAggregateFunctionExpression(GlobalAttribute attribute) {
+        super(attribute);
     }
 
-    public MaxAggregateFunctionExpression(String rootFeatureName, String attributeName) {
-        super(rootFeatureName, attributeName);
+    public MaxAggregateFunctionExpression(GlobalAttribute attribute, Feature rootFeature) {
+        super(attribute, rootFeature);
     }
 
     @Override
