@@ -1,19 +1,26 @@
 package de.vill.model.expression;
 
 import de.vill.model.Feature;
+import de.vill.model.building.VariableReference;
 import de.vill.util.Constants;
-import java.util.Arrays;
-import java.util.List;
-import java.util.Objects;
-import java.util.Set;
 
-public class SubExpression extends Expression {
+import java.util.*;
+
+public class SubExpression extends BinaryExpression {
     private Expression left;
     private Expression right;
 
     public SubExpression(Expression left, Expression right) {
         this.left = left;
         this.right = right;
+    }
+
+    public Expression getLeft() {
+        return left;
+    }
+
+    public Expression getRight() {
+        return right;
     }
 
     @Override
@@ -39,20 +46,7 @@ public class SubExpression extends Expression {
 
     @Override
     public double evaluate(Set<Feature> selectedFeatures) {
-        double leftResult;
-        if (left instanceof LiteralExpression && !selectedFeatures.contains(((LiteralExpression) left).getFeature())) {
-            leftResult = 0;
-        } else {
-            leftResult = left.evaluate(selectedFeatures);
-        }
-        double rightResult;
-        if (right instanceof LiteralExpression
-            && !selectedFeatures.contains(((LiteralExpression) right).getFeature())) {
-            rightResult = 0;
-        } else {
-            rightResult = right.evaluate(selectedFeatures);
-        }
-        return leftResult + rightResult;
+        return left.evaluate(selectedFeatures) - right.evaluate(selectedFeatures);
     }
 
     @Override
@@ -73,6 +67,13 @@ public class SubExpression extends Expression {
         }
         SubExpression other = (SubExpression) obj;
         return Objects.equals(left, other.left) && Objects.equals(right, other.right);
+    }
+
+    @Override
+    public List<VariableReference> getReferences() {
+        List<VariableReference> result = new ArrayList<>(left.getReferences());
+        result.addAll(right.getReferences());
+        return result;
     }
 
     @Override
