@@ -1,16 +1,19 @@
 package de.vill.model.expression;
 
+import de.vill.model.Attribute;
+import de.vill.model.Feature;
+import de.vill.model.GlobalAttribute;
 import de.vill.util.Constants;
 import java.util.Collections;
 import java.util.List;
 
 public class SumAggregateFunctionExpression extends AggregateFunctionExpression {
-    public SumAggregateFunctionExpression(String attributeName) {
-        super(attributeName);
+    public SumAggregateFunctionExpression(GlobalAttribute attribute) {
+        super(attribute);
     }
 
-    public SumAggregateFunctionExpression(String rootFeatureName, String attributeName) {
-        super(rootFeatureName, attributeName);
+    public SumAggregateFunctionExpression(GlobalAttribute attribute, Feature rootFeature) {
+        super(attribute, rootFeature);
     }
 
     @Override
